Day 9 / clean solution

+++ AdventOfCode2019.Core/Day9.cs
@@ -8,21 +8,20 @@ namespace AdventOfCode2019.Core
 {
     public class Day9
     {
-        public List<float> GetBoostCode(string path, int? input)
+        public List<decimal> GetBoostCode(string path, int? input)
         {
             Day7InputReader day2InputReader = new Day7InputReader();
-            List<float> fileContent = day2InputReader.ReadArray(path, ',').ToList();
+            List<decimal> fileContent = day2InputReader.ReadArray(path, ',').ToList();
 
             Amplifier amp = new Amplifier();
             amp.Memory = fileContent;
-            amp.Input = new List<float>();
+            amp.Input = new List<decimal>();
             if (input.HasValue)
             {
                 amp.Input.Add(input.Value);
             }
 
-            amp.Output = new List<float>();
-
+            amp.Output = new List<decimal>();
             while (!amp.Finished)
             {
                 amp.IntCode();
